requirements.txt

python-dotenv
sounddevice
numpy
websockets
requests
pydantic
fastapi
uvicorn
